requirements.txt
cmudict
cn2an
g2p_en
GPUtil
jieba
loguru
uvicorn
num2words
numpy<2
psutil
fastapi
torch<2.4
pyannote.audio>=3.1.0
pyopenjtalk-dict
pypinyin
pyworld-prebuilt
transformers==4.39.3
umap-learn
imagehash
Pillow
opencv-python-headless
paddlepaddle==3.0.0 
paddleocr>=2.7.3
googletrans-py
requests
sounddevice
soundfile
tkinter-async-execute
